Abstract away new helper functions and reorganise existing helper functions
