+++ src/curated-corpus/components/ScheduleItemFormConnector/ScheduleItemFormConnector.test.tsx
@@ -178,8 +178,8 @@ describe('ScheduleItemFormConnector', () => {
 
     // Click on the date to fire off the query to look up "already scheduled for this day" counts.
     // It's the second "date" pretend button, as the first one is actually hidden from view.
-    // Note: doesn't work!
-    userEvent.click(dates[1]);
+    // Note: doesn't work! ...and is super flaky!
+    //userEvent.click(dates[1]);
 
     // TODO:
     //  - ????
