+++ src/curated-corpus/helpers/publishers.ts
@@ -1,0 +1,30 @@
+import { GetScheduledItemsQuery } from '../../api/generatedTypes';
+import { ScheduleSummary } from '../components';
+
+/**
+ * Extracts a summary of publisher data from scheduled corpus item data.
+ *
+ * @param data
+ */
+export const getGroupedPublisherData = (
+  data: GetScheduledItemsQuery
+): ScheduleSummary[] => {
+  const publishers: ScheduleSummary[] = [];
+
+  data.getScheduledCorpusItems[0]?.items.forEach((item) => {
+    const publisher = item.approvedItem.publisher;
+
+    const existingEntry = publishers.find((entry) => entry.name === publisher);
+
+    existingEntry
+      ? (existingEntry.count += 1)
+      : publishers.push({ name: publisher, count: 1 });
+  });
+
+  // Sort publishers in descending order - most frequent on top
+  // And do a secondary sort within publishers with equal counts
+  // to list them in alphabetical order
+  publishers.sort((a, b) => b.count - a.count || a.name.localeCompare(b.name));
+
+  return publishers;
+};
