+++ src/curated-corpus/helpers/topics.ts
@@ -1,0 +1,61 @@
+import { GetScheduledItemsQuery, Maybe } from '../../api/generatedTypes';
+import {
+  DropdownOption,
+  topics as canonicalTopics,
+  topics,
+} from './definitions';
+import { ScheduleSummary } from '../components';
+
+/**
+ * This function transforms topic names as recorded in the database
+ * into more easily readable names, e.g. `TECHNOLOGY` -> `Technology`
+ * or `HEALTH_FITNESS` -> `Health & Fitness`.
+ *
+ * Returns `N/A` if there is no topic match from the known list of topics.
+ */
+export const getDisplayTopic = (
+  topicCode: Maybe<string> | string | undefined
+): string => {
+  const displayTopic = topics.find((topic) => {
+    return topic.code === topicCode;
+  })?.name;
+
+  return displayTopic ? displayTopic : 'N/A';
+};
+
+/**
+ * Extracts a summary of topic data from scheduled corpus item data.
+ *
+ * @param data
+ */
+export const getGroupedTopicData = (
+  data: GetScheduledItemsQuery
+): ScheduleSummary[] => {
+  const topics: ScheduleSummary[] = [];
+
+  data.getScheduledCorpusItems[0]?.items.forEach((item) => {
+    const topicName = getDisplayTopic(item.approvedItem.topic);
+
+    const existingEntry = topics.find((entry) => entry.name === topicName);
+
+    existingEntry
+      ? (existingEntry.count += 1)
+      : topics.push({ name: topicName, count: 1 });
+  });
+
+  // Add the rest of the pre-defined topics - we need to list them all
+  canonicalTopics.forEach((topic: DropdownOption) => {
+    const topicExists = topics.find((entry) => entry.name === topic.name);
+
+    if (!topicExists) {
+      topics.push({ name: topic.name, count: 0 });
+    }
+  });
+
+  // Sort topics in descending order - most frequent on top
+  // And do a secondary sort within topics with equal counts
+  // to list them in alphabetical order
+  topics.sort((a, b) => b.count - a.count || a.name.localeCompare(b.name));
+
+  return topics;
+};
